Added new fields to product group data files.  - Old files should NOT be used in this version  - Statistics are removed on application close  - Application close is not instant as it waits for data thread to finalize

+++ src/com/ui/AddItemDialog.java
@@ -42,7 +42,7 @@ public class AddItemDialog extends JDialog {
                             if (StringRegExChecker.checkInteger(productCountTextField.getText())) {
                                 parentFrame.getCurrentGroup().add(new Product(newProductName, productDescTextPane.getText(),
                                         productManufacturerTextField.getText(), Integer.parseInt(productCountTextField.getText()),
-                                        Double.parseDouble(productPriceTextField.getText())));
+                                        Float.parseFloat(productPriceTextField.getText())));
                                 dispose();
                                 cache.set(parentFrame.getCurrentGroup());
                                 parentFrame.reload();
